bfloat16 -> float16

+++ interactivechat.py
@@ -15,7 +15,7 @@ tokenizer = vl_chat_processor.tokenizer
 vl_gpt: MultiModalityCausalLM = AutoModelForCausalLM.from_pretrained(
     model_path, trust_remote_code=True
 )
-vl_gpt = vl_gpt.to(torch.bfloat16).cuda().eval()
+vl_gpt = vl_gpt.to(torch.float16).cuda().eval()
 
 
 def create_prompt(user_input: str) -> str:
