+++ janus/utils/io.py
@@ -36,7 +36,7 @@ def load_pretrained_model(model_path: str):
     vl_gpt: MultiModalityCausalLM = AutoModelForCausalLM.from_pretrained(
         model_path, trust_remote_code=True
     )
-    vl_gpt = vl_gpt.to(torch.bfloat16).cuda().eval()
+    vl_gpt = vl_gpt.to(torch.float16).cuda().eval()
 
     return tokenizer, vl_chat_processor, vl_gpt
